requirements.txt
# Airflow
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.0.0

# Dependencias necesarias
idna==3.7
lxml==5.2.2
multitasking==0.0.11
#numpy==1.23.5
packaging==24.0
#pandas==2.0.2
peewee==3.17.5
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==1.4.52
sqlalchemy-redshift==0.8.14
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
yfinance==0.2.40
#psycopg2==2.9.9
#pendulum==2.1.2